Revert "refactor: extract utils from different layers"
This reverts commit da7585a3b7a9d554ed20688b63769498757d6dc5.
