+++ app/src/main/java/io/github/chhabra_dhiraj/spaceflightnews/feature/article/data/mapper/ArticleMapper.kt
@@ -2,7 +2,7 @@ package io.github.chhabra_dhiraj.spaceflightnews.feature.article.data.mapper
 
 import io.github.chhabra_dhiraj.spaceflightnews.feature.article.data.remote.ArticleDto
 import io.github.chhabra_dhiraj.spaceflightnews.feature.article.data.remote.ArticleListDto
-import io.github.chhabra_dhiraj.spaceflightnews.util.datetime.getLocalDateTime
+import io.github.chhabra_dhiraj.spaceflightnews.feature.article.data.util.getLocalDateTime
 import io.github.chhabra_dhiraj.spaceflightnews.feature.article.domain.model.Article
 
 fun ArticleDto.toArticle() = Article(
