+++ app/src/main/java/io/github/chhabra_dhiraj/spaceflightnews/feature/article/domain/util/Error.kt
@@ -1,0 +1,3 @@
+package io.github.chhabra_dhiraj.spaceflightnews.feature.article.domain.util
+
+sealed interface Error
