+++ app/src/main/java/io/github/chhabra_dhiraj/spaceflightnews/feature/article/presentation/articledetail/BaseArticleDetailScreen.kt
@@ -13,7 +13,7 @@ import androidx.navigation.NavController
 import androidx.navigation.toRoute
 import io.github.chhabra_dhiraj.spaceflightnews.ArticleDetail
 import io.github.chhabra_dhiraj.spaceflightnews.NavigationEvent
-import io.github.chhabra_dhiraj.spaceflightnews.util.compose.rememberFlowWithLifecycle
+import io.github.chhabra_dhiraj.spaceflightnews.feature.article.presentation.util.rememberFlowWithLifecycle
 
 @Composable
 fun BaseArticleDetailScreen(
