+++ app/src/main/java/io/github/chhabra_dhiraj/spaceflightnews/feature/article/presentation/articlelist/BaseArticleListScreen.kt
@@ -8,7 +8,7 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.NavController
 import io.github.chhabra_dhiraj.spaceflightnews.ArticleDetail
 import io.github.chhabra_dhiraj.spaceflightnews.NavigationEvent
-import io.github.chhabra_dhiraj.spaceflightnews.util.compose.rememberFlowWithLifecycle
+import io.github.chhabra_dhiraj.spaceflightnews.feature.article.presentation.util.rememberFlowWithLifecycle
 
 @Composable
 fun BaseArticleListScreen(
