+++ app/src/main/java/io/github/chhabra_dhiraj/spaceflightnews/feature/article/presentation/articlelist/component/ArticleListItem.kt
@@ -20,7 +20,7 @@ import io.github.chhabra_dhiraj.spaceflightnews.feature.article.domain.model.Art
 import io.github.chhabra_dhiraj.spaceflightnews.feature.article.domain.sampledata.getSampleArticleList
 import io.github.chhabra_dhiraj.spaceflightnews.feature.article.presentation.component.ImageArticle
 import io.github.chhabra_dhiraj.spaceflightnews.feature.article.presentation.ui.theme.SpaceflightNewsTheme
-import io.github.chhabra_dhiraj.spaceflightnews.util.datetime.getFullDate
+import io.github.chhabra_dhiraj.spaceflightnews.feature.article.presentation.util.getFullDate
 
 @Composable
 fun ArticleListItem(
