+++ app/src/main/java/io/github/chhabra_dhiraj/spaceflightnews/feature/article/presentation/util/LocalDateTimeExt.kt
@@ -1,4 +1,4 @@
-package io.github.chhabra_dhiraj.spaceflightnews.util.datetime
+package io.github.chhabra_dhiraj.spaceflightnews.feature.article.presentation.util
 
 import java.time.LocalDateTime
 import java.time.ZoneId
