--- app/src/main/java/io/github/chhabra_dhiraj/spaceflightnews/util/network/Error.kt
@@ -1,3 +1,0 @@
-package io.github.chhabra_dhiraj.spaceflightnews.util.network
-
-sealed interface Error
